gcc 3.2 changes
git-svn-id: https://workrave.svn.sourceforge.net/svnroot/workrave/workrave/workrave/trunk@131 2b0e655b-6f86-497a-824e-95ef0e1e0d90

+++ gui/gtkmm/BreakWindow.cc
@@ -201,7 +201,7 @@ BreakWindow::set_avoid_pointer(bool avoid_pointer)
 bool
 BreakWindow::on_enter_notify_event(GdkEventCrossing *event)
 {
-  avoid_pointer(event->x, event->y);
+  avoid_pointer((int)event->x, (int)event->y);
   return Gtk::Window::on_enter_notify_event(event);
 }
 #endif
